FileSaveDialog: Inter-Object interaction working
